HW06: 2.1 Добавить транзакционность (DataSourceTransactionManager) в Jdbc-реализации 2.2 Добавить еще одну роль к юзеру Admin (будет 2 роли: ROLE_USER, ROLE_ADMIN). 2.3 В JdbcUserRepository добавить реализацию ролей юзера (добавлять можно одним запросом с JOIN и RowMapper, либо двумя запросами (отдельно users и отдельно roles). Объяснение SQL JOIN 2.3.1 В реализации getAll НЕ делать запрос ролей для каждого юзера (N+1 select) 2.3.2 При save посмотрите на batchUpdate() 2.4 Добавить проверку ролей в UserTestData.USER_MATCHER.assertMatch и починить ВСЕ тесты (тесты должны проходить для юзера с несколькими ролями)

+++ src/main/resources/db/populateDB.sql
@@ -1,6 +1,6 @@
-DELETE FROM user_roles;
-DELETE FROM meals;
-DELETE FROM users;
+DELETE FROM user_roles WHERE user_id > 1;
+DELETE FROM meals WHERE id > 1;
+DELETE FROM users WHERE id > 1;
 ALTER SEQUENCE global_seq RESTART WITH 100000;
 
 INSERT INTO users (name, email, password) VALUES
@@ -9,7 +9,8 @@ INSERT INTO users (name, email, password) VALUES
 
 INSERT INTO user_roles (role, user_id) VALUES
   ('ROLE_USER', 100000),
-  ('ROLE_ADMIN', 100001);
+  ('ROLE_ADMIN', 100001),
+  ('ROLE_USER', 100001);
 
 INSERT INTO meals (date_time, description, calories, user_id)
 VALUES ('2020-01-30 10:00:00', 'Завтрак', 500, 100000),
